Maze
I created a small maze and trained the agent in it. Tweaking the values a bit helped it learn to navigate it, though the raycasts should either be shortened or the distance being added back in.

+++ Assets/Scripts/AgentExampleControls.cs
@@ -7,7 +7,7 @@ public class AgentExampleControls : MonoBehaviour {
     [SerializeField] private BasicANNInitializer ANNInitializer = null;
     [SerializeField] private SOANNData trainingData;
     [SerializeField] private SOANNData liveData;
-    [SerializeField] private int rayLength = 5;
+    [SerializeField] private float rayLength = 5;
     [SerializeField] private GameObject ANNView = null;
     private bool isRecievingInput = false;
     private double forward = 0;
